Precheck runtime, To #43759509
