--- pkg/runtime/alluxio/precheck.go
@@ -1,32 +1,0 @@
-/*
-Copyright 2022 The Fluid Author.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-
-*/
-
-package alluxio
-
-import (
-	"k8s.io/apimachinery/pkg/types"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-
-	datav1alpha1 "github.com/fluid-cloudnative/fluid/api/v1alpha1"
-	"github.com/fluid-cloudnative/fluid/pkg/utils"
-)
-
-// Precheck checks if the given key can be found in the current runtime types
-func Precheck(client client.Client, key types.NamespacedName) (found bool, err error) {
-	var obj datav1alpha1.AlluxioRuntime
-	return utils.CheckObject(client, key, &obj)
-}
