+++ pkg/runtime/runtime_controllers.go
@@ -23,10 +23,6 @@ import (
 	"strconv"
 
 	"github.com/fluid-cloudnative/fluid/pkg/common"
-	"github.com/fluid-cloudnative/fluid/pkg/runtime/alluxio"
-	"github.com/fluid-cloudnative/fluid/pkg/runtime/goosefs"
-	"github.com/fluid-cloudnative/fluid/pkg/runtime/jindofsx"
-	"github.com/fluid-cloudnative/fluid/pkg/runtime/juicefs"
 	"github.com/go-logr/logr"
 	appsv1 "k8s.io/api/apps/v1"
 	"k8s.io/apimachinery/pkg/types"
@@ -36,19 +32,18 @@ import (
 	utilpointer "k8s.io/utils/pointer"
 )
 
-type checkFunc func(client.Client, types.NamespacedName) (bool, error)
+type CheckFunc func(client.Client, types.NamespacedName) (bool, error)
 
-var checkFuncs map[string]checkFunc = map[string]checkFunc{
-	"alluxioruntime-controller": alluxio.Precheck,
-	"jindoruntime-controller":   jindofsx.Precheck,
-	"juicefsruntime-controller": juicefs.Precheck,
-	"goosefsruntime-controller": goosefs.Precheck,
+var precheckFuncs map[string]CheckFunc
+
+func SetPrecheckFunc(checks map[string]CheckFunc) {
+	precheckFuncs = checks
 }
 
 func ScaleoutRuntimeContollerOnDemand(c client.Client, datasetKey types.NamespacedName, log logr.Logger) (
 	controllerName string, scaleout bool, err error) {
 
-	for myControllerName, checkRuntime := range checkFuncs {
+	for myControllerName, checkRuntime := range precheckFuncs {
 		match, err := checkRuntime(c, datasetKey)
 		if err != nil {
 			return controllerName, scaleout, err
